+++ src/main/java/InstructionType.java
@@ -1,0 +1,12 @@
+public enum InstructionType {
+    NONE,
+    UNKNOWN,
+    EXIT,
+    ADD_TODO,
+    ADD_EVENT,
+    ADD_DEADLINE,
+    PRINT_LIST,
+    MARK_ITEM,
+    UNMARK_ITEM,
+    DELETE_ITEM
+}
