+++ src/main/java/ToDo.java
@@ -1,0 +1,5 @@
+public class ToDo extends Task {
+    public ToDo(String[] args) throws DukeException {
+        super("todo", args);
+    }
+}
